refactor: moved the exact scalar wire form next to PdbOwnedValue.
The tagged round-trip form is a property of the value, not the KD-tree.
It sits with `total_cmp` and the lossy default serde, is reusable, and
keeps `kdtree.rs` off the Postgres datetime types.
