+++ pg_search/src/index/kdtree.rs
@@ -66,75 +66,11 @@ enum KdNode {
     },
     Split {
         dim: usize,
-        #[serde(with = "wire_value")]
+        #[serde(with = "crate::postgres::pdb_owned_value::exact_scalar_wire")]
         value: PdbOwnedValue,
         left: Box<KdNode>,
         right: Box<KdNode>,
     },
-}
-
-/// `PdbOwnedValue`'s own serde form goes through tantivy's untagged `OwnedValue`, which turns
-/// every non-negative `I64` into a `U64` on the way back. A split value that changed variant
-/// would no longer compare like the column it came from, so split values travel tagged.
-///
-/// Dates travel as their raw microsecond count. Their string form goes through Postgres's
-/// datetime output, which the tree must not depend on: it is also built and tested outside a
-/// backend.
-mod wire_value {
-    use std::net::Ipv6Addr;
-
-    use serde::{Deserialize, Deserializer, Serialize, Serializer, de, ser};
-
-    use crate::postgres::datetime::PostgresDateTime;
-    use crate::postgres::pdb_owned_value::PdbOwnedValue;
-
-    #[derive(Serialize, Deserialize)]
-    enum WireValue {
-        Str(String),
-        U64(u64),
-        I64(i64),
-        Bool(bool),
-        Date(i64),
-        // `f64::to_bits`, so `NaN` and the infinities survive any codec, text or binary.
-        F64Bits(u64),
-        Bytes(Vec<u8>),
-        IpAddr(Ipv6Addr),
-    }
-
-    pub fn serialize<S: Serializer>(value: &PdbOwnedValue, s: S) -> Result<S::Ok, S::Error> {
-        let wire = match value {
-            PdbOwnedValue::Str(v) => WireValue::Str(v.clone()),
-            PdbOwnedValue::U64(v) => WireValue::U64(*v),
-            PdbOwnedValue::I64(v) => WireValue::I64(*v),
-            PdbOwnedValue::F64(v) => WireValue::F64Bits(v.to_bits()),
-            PdbOwnedValue::Bool(v) => WireValue::Bool(*v),
-            PdbOwnedValue::Date(v) => WireValue::Date(v.into_inner()),
-            PdbOwnedValue::Bytes(v) => WireValue::Bytes(v.clone()),
-            PdbOwnedValue::IpAddr(v) => WireValue::IpAddr(*v),
-            other => {
-                return Err(ser::Error::custom(format!(
-                    "unsupported partition split value: {other:?}"
-                )));
-            }
-        };
-        wire.serialize(s)
-    }
-
-    pub fn deserialize<'de, D: Deserializer<'de>>(d: D) -> Result<PdbOwnedValue, D::Error> {
-        Ok(match WireValue::deserialize(d)? {
-            WireValue::Str(v) => PdbOwnedValue::Str(v),
-            WireValue::U64(v) => PdbOwnedValue::U64(v),
-            WireValue::I64(v) => PdbOwnedValue::I64(v),
-            WireValue::F64Bits(v) => PdbOwnedValue::F64(f64::from_bits(v)),
-            WireValue::Bool(v) => PdbOwnedValue::Bool(v),
-            WireValue::Date(v) => PdbOwnedValue::Date(
-                PostgresDateTime::try_from_raw(v)
-                    .map_err(|e| de::Error::custom(format!("invalid split date: {e:?}")))?,
-            ),
-            WireValue::Bytes(v) => PdbOwnedValue::Bytes(v),
-            WireValue::IpAddr(v) => PdbOwnedValue::IpAddr(v),
-        })
-    }
 }
 
 /// The half-open range a partition covers on one dimension.
